get genes using data portal index.
